Fix ensure_build_test
Isolate it in its own file that we run separately on travis

Fixes https://github.com/dart-lang/json_serializable/issues/55
