+++ json_serializable/test/ensure_build_test.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 @TestOn('vm')
+@Tags(const ['presubmit-only'])
 
 import 'dart:convert';
 import 'dart:io';
@@ -10,48 +11,7 @@ import 'dart:io';
 import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
 
-String _getExampleContent(String fileName) {
-  var lines = new File(p.join('example', fileName)).readAsLinesSync();
-
-  var lastHadContent = false;
-
-  // All lines with content, except those starting with `/`.
-  // Also exclude blank lines that follow other blank lines
-  var cleanedSource = lines.where((l) {
-    if (l.startsWith(r'/')) {
-      return false;
-    }
-
-    if (l.trim().isNotEmpty) {
-      lastHadContent = true;
-      return true;
-    }
-
-    if (lastHadContent) {
-      lastHadContent = false;
-      return true;
-    }
-
-    return false;
-  }).join('\n');
-
-  return '''
-```dart
-$cleanedSource
-```''';
-}
-
 void main() {
-  test('README example', () {
-    var readmeContent = new File('README.md').readAsStringSync();
-
-    var exampleContent = _getExampleContent('example.dart');
-    expect(readmeContent, contains(exampleContent));
-
-    var exampleGeneratedContent = _getExampleContent('example.g.dart');
-    expect(readmeContent, contains(exampleGeneratedContent));
-  });
-
   // TODO(kevmoo): replace with a common utility
   //               https://github.com/dart-lang/build/issues/716
   test('ensure local build succeeds with no changes', () {
@@ -67,14 +27,16 @@ void main() {
     expect(_changedGeneratedFiles(), isEmpty);
 
     // 2 - run build - should be no output, since nothing should change
-    var result = _runProc('pub',
-        ['run', 'build_runner', 'build', '--delete-conflicting-outputs']);
+    var result = _runProc(
+        'dart', ['tool/build.dart', 'build', '--delete-conflicting-outputs']);
+
+    print(result);
     expect(result,
         contains(new RegExp(r'\[INFO\] Succeeded after \S+ with \d+ outputs')));
 
     // 3 - get a list of modified `.g.dart` files - should still be empty
     expect(_changedGeneratedFiles(), isEmpty);
-  }, tags: 'presubmit-only');
+  });
 }
 
 final _whitespace = new RegExp(r'\s');
