+++ tool/travis.sh
@@ -34,10 +34,20 @@ while (( "$#" )); do
     echo -e 'dartfmt -n --set-exit-if-changed .'
     dartfmt -n --set-exit-if-changed . || EXIT_CODE=$?
     ;;
-  test) echo
-    echo -e '\033[1mTASK: test\033[22m'
+  test_0) echo
+    echo -e '\033[1mTASK: test_0\033[22m'
     echo -e 'pub run test --run-skipped'
     pub run test --run-skipped || EXIT_CODE=$?
+    ;;
+  test_1) echo
+    echo -e '\033[1mTASK: test_1\033[22m'
+    echo -e 'pub run test'
+    pub run test || EXIT_CODE=$?
+    ;;
+  test_2) echo
+    echo -e '\033[1mTASK: test_2\033[22m'
+    echo -e 'pub run test --run-skipped test/ensure_build_test.dart'
+    pub run test --run-skipped test/ensure_build_test.dart || EXIT_CODE=$?
     ;;
   *) echo -e "\033[31mNot expecting TASK '${TASK}'. Error!\033[0m"
     EXIT_CODE=1
